requirements.txt
supabase>=2.3.1
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp
crawl4ai
boto3
asyncio
playwright>=1.40.0
httpx>=0.25.2,<0.26.0
ollama==0.1.0
celery>=5.3.0
redis>=5.0.0